+++ packages/s3-credentials-endpoint/index.js
@@ -24,10 +24,9 @@ const urljoin = require('url-join');
 const { AccessToken } = require('@cumulus/api/models');
 const { isLocalApi } = require('@cumulus/api/lib/testUtils');
 const { isAccessTokenExpired } = require('@cumulus/api/lib/token');
+const { handleCredentialRequest } = require('@cumulus/api/endpoints/s3credentials');
 const awsServices = require('@cumulus/aws-client/services');
 const { RecordDoesNotExist } = require('@cumulus/errors');
-
-const log = new Logger({ sender: 's3credentials' });
 
 // From https://docs.aws.amazon.com/STS/latest/APIReference/API_AssumeRole.html#API_AssumeRole_RequestParameters
 const roleSessionNameRegex = /^[\w+,.=@-]{2,64}$/;
@@ -90,43 +89,6 @@ async function displayS3CredentialInstructions(_req, res) {
   const instructionTemplate = await readFile(pathjoin(__dirname, 'instructions', 'index.html'), 'utf-8');
   const compiled = template(instructionTemplate);
   res.send(compiled(process.env));
-}
-
-/**
- * Dispenses time-based temporary credentials for same-region direct s3 access.
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- * @returns {Object} the express response object with object containing
- *                   tempoary s3 credentials for direct same-region s3 access.
- */
-async function s3credentials(req, res) {
-  const disableS3Credentials = process.env.DISABLE_S3_CREDENTIALS;
-
-  if (disableS3Credentials && (disableS3Credentials.toLowerCase() === 'true')) {
-    return res.boom.serverUnavailable('S3 Credentials Endpoint has been disabled');
-  }
-
-  const roleSessionName = buildRoleSessionName(
-    req.authorizedMetadata.userName,
-    req.authorizedMetadata.clientName
-  );
-
-  const credentials = await requestTemporaryCredentialsFromNgap({
-    lambda: req.lambda,
-    lambdaFunctionName: process.env.STSCredentialsLambda,
-    userId: req.authorizedMetadata.userName,
-    roleSessionName,
-  });
-
-  const creds = JSON.parse(credentials.Payload);
-  if (Object.keys(creds).some((key) => ['errorMessage', 'errorType', 'stackTrace'].includes(key))) {
-    log.error(credentials.Payload);
-    return res.boom.failedDependency(
-      `Unable to retrieve credentials from Server: ${credentials.Payload}`
-    );
-  }
-  return res.send(creds);
 }
 
 // Running API locally will be on http, not https, so cookies
@@ -192,19 +154,6 @@ async function handleRedirectRequest(req, res) {
     .set({ Location: urljoin(distributionUrl, state) })
     .status(307)
     .send('Redirecting');
-}
-
-/**
- * Responds to a request for temporary s3 credentials.
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- * @returns {Promise<Object>} the promise of express response object containing
- * temporary credentials
- */
-async function handleCredentialRequest(req, res) {
-  req.lambda = awsServices.lambda();
-  return s3credentials(req, res);
 }
 
 /**
@@ -375,6 +324,5 @@ module.exports = {
   distributionApp,
   handler,
   handleTokenAuthRequest,
-  requestTemporaryCredentialsFromNgap,
-  s3credentials,
+  requestTemporaryCredentialsFromNgap
 };